[CAP-20] Implement manage profile for attendant - Implement Mapper DTO to Entity without null field
