+++ src/main/java/org/capstone/job_fair/services/mappers/AttendantEntityMapper.java
@@ -8,9 +8,7 @@ import org.mapstruct.Mapper;
 
 public abstract class AttendantEntityMapper {
 
-
     public abstract AttendantDTO toDTO(AttendantEntity attendant);
-
     public abstract AttendantEntity toEntity(AttendantDTO dto);
 
 }
